requirement.txt
torch>=2.0.0
torchvision>=0.15.0
numpy<2.0
albumentations>=1.3.0
Pillow>=9.0.0
opencv-python>=4.5.0
tqdm>=4.60.0
matplotlib>=3.5.0
scipy>=1.7.0
